+++ R/dhs.R
@@ -1,4 +1,4 @@
-#' Read and recode UN Demographic and Health Surveys (DHS) individual data
+#' Read and recode Demographic and Health Surveys (DHS) individual data
 #'
 #' @param files vector: a character vector containing the paths for one or more Individual Recode DHS data files (see details)
 #' @param extra.vars vector: a character vector containing the names of variables to be retained from the raw data
@@ -11,13 +11,28 @@
 #'    files are available in SPSS, SAS, and Stata formats from \href{https://www.dhsprogram.com/}{https://www.dhsprogram.com/},
 #'    however access requires a \href{https://dhsprogram.com/data/Access-Instructions.cfm}{free application}. The `dhs()` function
 #'    reads one or more of these files, extracts and recodes selected variables useful for studying childfree adults and other
-#'    family statuses, then returns a single data frame.
+#'    family statuses, then returns an unweighted data frame.
 #'
-#' Although access to DHS data requires an application, the DHS program provides \href{https://dhsprogram.com/data/Download-Model-Datasets.cfm}{model datasets}
+#' Although access to DHS data requires an application, the DHS program provides a \href{https://dhsprogram.com/data/Download-Model-Datasets.cfm}{model dataset}
 #'    for practice. The example provided below uses the model data file "ZZIR62FL.SAV", which contains fictitious women's data,
-#'    but has the same structure as real DHS data files. The example can be run without prior application for data access.
+#'    but has the same structure as a real DHS data file. The example can be run without prior application for data access.
 #'
-#' **Known issues**
+#' **Sampling weights**
+#'
+#' The DHS is collected using a complex survey design. The \code{survey} package can be used to perform analyses that take these
+#'    design features into account, and make it possible to obtain population-representative estimates. In most cases, a \link[survey]{svydesign}
+#'    object for a single country and wave can be created using \code{survey::svydesign(data = data, ids = ~cluster, strata = ~strata, weights = ~weight, nest = TRUE)}.
+#'    Additional information about analyzing DHS data using weights is available \href{https://dhsprogram.com/data/Guide-to-DHS-Statistics/Analyzing_DHS_Data.htm}{here}
+#'    and in the documentation provided with the downloaded data files.
+#'
+#' **Non-biological children**
+#'
+#' Information about non-biological children (e.g., adopted children, foster children, etc.) is not available in the DHS, which means
+#'    that a respondent with only non-biological children would be classified as a non-parent. This is not exactly match the approach
+#'    described by the ABC Framework (Neal & Neal, 2024), and may lead to discrepancies when comparing DHS estimates to estimates derived
+#'    from other data where information about non-biological children is available.
+#'
+#' **Additional notes**
 #'   * The SPSS-formatted files containing data from Gabon Recode 4 (GAIR41FL.SAV, GAMR41FL.SAV) and Turkey Recode 4 (TRIR41FL.SAV, TRMR41FL.SAV)
 #'     contain encoding errors. Use the SAS-formatted files (GAIR41FL.SAS7BDAT, GAMR41FL.SAS7BDAT, TRIR41FL.SAS7BDAT, TRMR41FL.SAS7BDAT) instead.
 #'   * In some cases, DHS makes available individual recode data files for specific regions. For example, women's data from individual states
@@ -28,11 +43,18 @@
 #'     extra variables are requested using the `extra.vars` option, be sure to specify both names (e.g. `extra.vars = c("v201", "mv201")`).
 #'
 #' @return A data frame containing variables described in the codebook available using \code{vignette("codebooks")}
+#' If you are offline, or if the requested data are otherwise unavailable, NULL is returned.
 #'
+#' @references ABC Framework: {Neal, Z. P. and Neal, J. W. (2024). A framework for studying adults who neither have nor want children. *The Family Journal, 32*, 121-130. \doi{10.1177/10664807231198869}}
 #' @export
 #'
 #' @examples
-#' \donttest{data <- dhs(files = c("ZZIR62FL.SAV"), extra.vars = c("v201"))}
+#' \donttest{
+#' dat <- dhs(files = c("ZZIR62FL.SAV"), extra.vars = c("v201"))  #Request data for fictitous country
+#' if (!is.null(dat)) {  #If data was available...
+#' table(dat$famstat)/nrow(dat)  #Fraction of respondents with each family status
+#' }
+#' }
 dhs <- function(files, extra.vars = NULL, progress = TRUE) {
 
   if (length(files) > 1 & "ZZIR62FL.SAV" %in% files) {stop("Model data (file ZZIR62FL.SAV) should not be combined with files containing real data.")}
@@ -45,15 +67,16 @@ dhs <- function(files, extra.vars = NULL, progress = TRUE) {
   #Loop over each supplied data file
   for (file in 1:length(files)) {
 
-    #Increment progress bar
-    if (progress) {utils::setTxtProgressBar(pb,file)}
-
     #Import raw data
     if (files[file]=="ZZIR62FL.SAV") {  #Model file from https://dhsprogram.com/data/Download-Model-Datasets.cfm
+      if (!RCurl::url.exists("https://osf.io/download/hk23e")) {message("You are offline or sample DHS data is not available now. Try again later"); data <- NULL; return(data)}
       temp <- tempfile()
       utils::download.file(url = "https://osf.io/download/hk23e", destfile = temp)
       dat <- rio::import(temp, format = "sav")
-    } else {dat <- rio::import(files[file])}
+    } else {
+      if (progress) {utils::setTxtProgressBar(pb,file)}
+      dat <- rio::import(files[file])
+      }
     colnames(dat) <- tolower(colnames(dat))  #Make all variables lowercase
 
     #Check type of file
@@ -77,7 +100,7 @@ dhs <- function(files, extra.vars = NULL, progress = TRUE) {
     #Ideal number of children
     if (female) {dat$ideal <- dat$v613} else {dat$ideal <- dat$mv613}
     dat$ideal[which(dat$ideal==98)] <- -1  #Special code for "Don't Know"
-    dat$ideal[which(dat$ideal>30)] <- NA  #Treat all values above 30 as missing (includes some undocumented country-specific special codes)
+    dat$ideal[which(dat$ideal>=90)] <- NA  #Treat all 90s values as missing (undocumented country-specific special codes)
 
     #Childfree (want)
     dat$cf_want <- NA
@@ -255,15 +278,17 @@ dhs <- function(files, extra.vars = NULL, progress = TRUE) {
                        "KK", "KE", "KY", "LA", "LS", "LB", "MD", "MW", "MV", "ML", "MR", "MX", "MB", "MA", "MZ", "MM", "NM", "NP", "NC", "NI",
                        "NG", "OS", "PK", "PY", "PE", "PH", "RW", "WS", "ST", "SN", "SL", "ZA", "LK", "SD", "SZ", "TJ", "TZ", "TH", "TL", "TG",
                        "TT", "TN", "TR", "TM", "UG", "UA", "UZ", "VN", "YE", "ZM", "ZW", "PG", "ZZ")
-    country.names <- c("Afghanistan", "Albania", "Angola", "Armenia", "Azerbaijan", "Bangladesh", "Benin", "Bolivia", "Botswana", "Brazil", "Burkina Faso", "Burundi", "Cambodia", "Cameroon", "Cape Verde", "Central African Republic", "Chad", "Columbia", "Comoros", "Congo",
+    country.names <- c("Afghanistan", "Albania", "Angola", "Armenia", "Azerbaijan", "Bangladesh", "Benin", "Bolivia", "Botswana", "Brazil", "Burkina Faso", "Burundi", "Cambodia", "Cameroon", "Cape Verde", "Central African Republic", "Chad", "Colombia", "Comoros", "Congo",
                        "Congo Democratic Republic", "Cote d'Ivoire", "Dominican Republic", "Ecuador", "Egypt", "El Salvador", "Equatorial Guinea", "Eritrea", "Ethiopia", "Gabon", "Gambia", "Ghana", "Guatamala", "Guinea", "Guyana", "Haiti", "Honduras", "India", "Indonesia", "Jordan",
                        "Kazakhstan", "Kenya", "Kyrgyz Republic", "Lao People's Democratic Republic", "Lesotho", "Liberia", "Madagascar", "Malawi", "Maldives", "Mali", "Mauritania", "Mexico", "Moldova", "Morocco", "Mozambique", "Myanmar", "Namibia", "Nepal", "Nicaragua", "Niger",
                        "Nigeria", "Nigeria (Ondo State)", "Pakistan", "Paraguay", "Peru", "Philippines", "Rwanda", "Samoa", "Sao Tome and Principe", "Senegal", "Sierra Leone", "South Africa", "Sri Lanka", "Sudan", "Swaziland", "Tajikstan", "Tanzania", "Thailand", "Timor-Leste", "Togo",
                        "Trinidad and Tobago", "Tunisia", "Turkey", "Turkministan", "Uganda", "Ukraine", "Uzbekistan", "Vietnam", "Yemen", "Zambia", "Zimbabwe", "Papua New Guinea", "Fake Country")
     if (female) {dat$country <- country.names[match(substr(dat$v000,1,2), country.codes)]} else {dat$country <- country.names[match(substr(dat$mv000,1,2), country.codes)]}
 
-    #Sampling weight
+    #Weighting variables
     if (female) {dat$weight <- dat$v005/1000000} else {dat$weight <- dat$mv005/1000000}
+    if (female) {dat$cluster <- dat$v021} else {dat$cluster <- dat$mv021}
+    if (female) {dat$strata <- dat$v023} else {dat$strata <- dat$mv023}
 
     #Wave (called "Recode" in the DHS)
     if (female) {
@@ -309,13 +334,13 @@ dhs <- function(files, extra.vars = NULL, progress = TRUE) {
       dat <- dat[,c("cf_want", "cf_ideal", "famstat",  #Family status
                     "sex", "age", "education", "partnered", "residence", "employed",  #Demographics
                     "religion",  #Attitude
-                    "id", "country", "weight", "file", "survey", "wave", "year", "month",  #Design
+                    "id", "country", "weight", "cluster", "strata", "file", "survey", "wave", "year", "month",  #Design
                     extra.vars)]
     } else {
       dat <- dat[,c("cf_want", "cf_ideal", "famstat",  #Family status
                     "sex", "age", "education", "partnered", "residence", "employed",  #Demographics
                     "religion",  #Attitude
-                    "id", "country", "weight", "file", "survey", "wave", "year", "month")]  #Design
+                    "id", "country", "weight", "cluster", "strata", "file", "survey", "wave", "year", "month")]  #Design
     }
 
     #Start data file, or append to existing data file
@@ -326,5 +351,6 @@ dhs <- function(files, extra.vars = NULL, progress = TRUE) {
   #Finalize
   if (progress) {close(pb)}  #Close progress bar
   class(data) <- c("data.frame", "childfree")
-  return(data)  #Export data
+  return(data)
+
 }