+++ NEWS.md
@@ -3,6 +3,14 @@ title: "NEWS"
 output: html_document
 ---
 
+## childfree 0.0.4
+
+* replaced direct support for `survey` with documentation about how to use the `survey` package
+
+## childfree 0.0.3
+
+* added option to return an unweighted data.frame or weighted `survey` object
+
 ## childfree 0.0.2
 
 * added support for male respondents in `nsfg()` and `dhs()`